# docs/swagger.yaml
basePath: /
definitions:
...
  common.ApiErr:
    properties:
      Error:
        type: string
    type: object
...
  controls.resMarketStatus:
    properties:
      isOpening:
        type: boolean
      openTime:
        type: integer
    type: object
...
  services.CoinPriceView:
    properties:
      bigPrice:
        type: string
      coin:
        description: 目标币价
        type: string
      node:
        type: string
      price:
        type: number
      sign:
        description: Sign_Hash值由 Timestamp+Coin+VsCoin+BigPrice计算
        items:
          type: integer
        type: array
      timestamp:
        type: integer
      vsCoin:
        description: vs币价
        type: string
    type: object
...
  services.StockData:
    properties:
      Price:
        description: 平均价
        type: number
      Timestamp:
        description: unix 秒数
        type: integer
      avgSigns:
        description: 所有节点平均价格签名列表
        items:
          $ref: '#/definitions/services.StockNode'
        type: array
      bigPrice:
        description: Multiply the Price by 1000000000000000000 to remove decimals
        type: string
      code:
        description: 合约代码
        type: string
      dataType:
        description: 最高最低价１最高　２最低价
        type: integer
      isMarketOpening:
        type: boolean
      marketOpenTime:
        type: integer
      sign:
        description: 计算平均价格的节点的签名　Sign_Hash值由 Timestamp　DataType BigPrice Code计算
        items:
          type: integer
        type: array
      signs:
        description: 所有节点签名列表
        items:
          $ref: '#/definitions/services.StockNodeRaw'
        type: array
      stockCode:
        description: 股票代码
        type: string
    type: object
  services.StockNode:
    properties:
      Timestamp:
        description: unix 秒数
        type: integer
      bigPrice:
        description: Multiply the Price by 1000000000000000000 to remove decimals
        type: string
      code:
        description: 合约代码
        type: string
      dataType:
        description: 最高最低价１最高　２最低价
        type: integer
      msg:
        description: debug
        type: string
      node:
        description: 节点名字
        type: string
      nodeAddress:
        description: 节点地址
        type: string
      price:
        description: 新价
        type: number
      sign:
        description: Sign_Hash值由 Timestamp　DataType BigPrice Code计算
        items:
          type: integer
        type: array
      stockCode:
        description: 股票代码
        type: string
    type: object
  services.StockNodeRaw:
    properties:
      Timestamp:
        description: unix 秒数
        type: integer
      bigPrice:
        description: Multiply the Price by 1000000000000000000 to remove decimals
        type: string
      code:
        description: 合约代码
        type: string
      dataType:
        description: 最高最低价１最高　２最低价
        type: integer
      msg:
        description: debug
        type: string
      node:
        description: 节点名字
        type: string
      nodeAddress:
        description: 节点地址
        type: string
      price:
        description: 新价
        type: number
      sign:
        description: Sign_Hash值由 Timestamp　DataType BigPrice Code计算
        items:
          type: integer
        type: array
      stockCode:
        description: 股票代码
        type: string
    type: object
...
info:
  contact:
    email: wuxiaofei@rechaintech.com
    name: 伍晓飞
  description: oracle-api接口文档.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://test.com/index.html
  title: oracle-api
  version: "1.0"
paths:
  /pub/coin_price/{coin}/{vs_coin}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取币价换算，多节点签名版
      operationId: CoinPriceSignHandler
      parameters:
      - default: eth
        description: 目标币价
        enum:
        - btc
        - aed
        - ars
        - aud
        - bch
        - bdt
        - bhd
        - bits
        - bmd
        - bnb
        - brl
        - byn
        - cad
        - chf
        - clp
        - cny
        - czk
        - dkk
        - dot
        - eos
        - eth
        - eur
        - gbp
        - hkd
        - huf
        - idr
        - ils
        - inr
        - jpy
        - krw
        - kwd
        - link
        - lkr
        - ltc
        - mmk
        - mxn
        - myr
        - ngn
        - nok
        - nzd
        - php
        - pkr
        - pln
        - rub
        - sar
        - sats
        - sek
        - sgd
        - thb
        - try
        - twd
        - uah
        - usd
        - vef
        - vnd
        - xag
        - xau
        - xdr
        - xlm
        - xrp
        - yfi
        - zar
        in: path
        name: coin
        required: true
        type: string
      - default: usd
        description: vs币价
        enum:
        - btc
        - aed
        - ars
        - aud
        - bch
        - bdt
        - bhd
        - bits
        - bmd
        - bnb
        - brl
        - byn
        - cad
        - chf
        - clp
        - cny
        - czk
        - dkk
        - dot
        - eos
        - eth
        - eur
        - gbp
        - hkd
        - huf
        - idr
        - ils
        - inr
        - jpy
        - krw
        - kwd
        - link
        - lkr
        - ltc
        - mmk
        - mxn
        - myr
        - ngn
        - nok
        - nzd
        - php
        - pkr
        - pln
        - rub
        - sar
        - sats
        - sek
        - sgd
        - thb
        - try
        - twd
        - uah
        - usd
        - vef
        - vnd
        - xag
        - xau
        - xdr
        - xlm
        - xrp
        - yfi
        - zar
        in: path
        name: vs_coin
        required: true
        type: string
      - default: 1620383144
        description: unix 秒数
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: CoinPriceView
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            $ref: '#/definitions/services.CoinPriceView'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取币价换算，多节点签名版
      tags:
      - default
...
  /pub/internal/coin_price/{coin}/{vs_coin}:
    get:
      consumes:
      - application/json
      description: 获取币价换算，内部单节点
      operationId: CoinPriceHandler
      parameters:
      - default: eth
        description: 目标币价
        enum:
        - btc
        - aed
        - ars
        - aud
        - bch
        - bdt
        - bhd
        - bits
        - bmd
        - bnb
        - brl
        - byn
        - cad
        - chf
        - clp
        - cny
        - czk
        - dkk
        - dot
        - eos
        - eth
        - eur
        - gbp
        - hkd
        - huf
        - idr
        - ils
        - inr
        - jpy
        - krw
        - kwd
        - link
        - lkr
        - ltc
        - mmk
        - mxn
        - myr
        - ngn
        - nok
        - nzd
        - php
        - pkr
        - pln
        - rub
        - sar
        - sats
        - sek
        - sgd
        - thb
        - try
        - twd
        - uah
        - usd
        - vef
        - vnd
        - xag
        - xau
        - xdr
        - xlm
        - xrp
        - yfi
        - zar
        in: path
        name: coin
        required: true
        type: string
      - default: usd
        description: vs币价
        enum:
        - btc
        - aed
        - ars
        - aud
        - bch
        - bdt
        - bhd
        - bits
        - bmd
        - bnb
        - brl
        - byn
        - cad
        - chf
        - clp
        - cny
        - czk
        - dkk
        - dot
        - eos
        - eth
        - eur
        - gbp
        - hkd
        - huf
        - idr
        - ils
        - inr
        - jpy
        - krw
        - kwd
        - link
        - lkr
        - ltc
        - mmk
        - mxn
        - myr
        - ngn
        - nok
        - nzd
        - php
        - pkr
        - pln
        - rub
        - sar
        - sats
        - sek
        - sgd
        - thb
        - try
        - twd
        - uah
        - usd
        - vef
        - vnd
        - xag
        - xau
        - xdr
        - xlm
        - xrp
        - yfi
        - zar
        in: path
        name: vs_coin
        required: true
        type: string
      - default: 1620383144
        description: unix 秒数
        in: query
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Coin Price
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            $ref: '#/definitions/services.CoinPriceView'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取币价换算，内部单节点
      tags:
      - default
...
  /pub/internal/stock_avgprice:
    post:
      consumes:
      - application/json
      description: 获取股票平均价格共识 苹果代码  AAPL  ,苹果代码 TSLA
      operationId: StockAvgPriceHandler
      parameters:
      - description: 节点价格列表
        in: body
        name: nodePrices
        required: true
        schema:
          items:
            $ref: '#/definitions/services.StockNode'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: stock info
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            $ref: '#/definitions/services.StockNode'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: '获取股票平均价格共识:'
      tags:
      - default
...
  /pub/stock/aggre_info/{code}/{data_type}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取共识美股价格 苹果代码  AAPL  ,苹果代码 TSLA
      operationId: StockAggreHandler
      parameters:
      - default: AAPL
        description: 美股代码
        enum:
        - USD
        - AAPL
        - TSLA
        in: path
        name: code
        required: true
        type: string
      - default: 1
        description: 最高最低价１最高　２最低价
        enum:
        - 1
        - 2
        in: path
        name: data_type
        required: true
        type: integer
      - default: 1620383144
        description: unix 秒数
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: stock info list
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            $ref: '#/definitions/services.StockData'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: '获取共识美股价格:'
      tags:
      - default
...
  /pub/stock/info/{code}/{data_type}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取美股价格 苹果代码  AAPL  ,苹果代码 TSLA
      operationId: StockInfoHandler
      parameters:
      - default: AAPL
        description: 美股代码
        enum:
        - AAPL
        - TSLA
        - USD
        in: path
        name: code
        required: true
        type: string
      - default: 1
        description: 最高最低价１最高　２最低价
        enum:
        - 1
        - 2
        in: path
        name: data_type
        required: true
        type: integer
      - default: 1620383144
        description: unix 秒数
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: stock info
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            $ref: '#/definitions/services.StockNode'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: '获取美股价格:'
      tags:
      - default
  /pub/stock/market_status/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取美股市场开盘状态,支持节假日,夏令时
      operationId: UsaMarketStatusHandler
      parameters:
      - default: 0
        description: unix 秒数； 0表示当前时间
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            $ref: '#/definitions/controls.resMarketStatus'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: '获取美股市场开盘状态:'
      tags:
      - default
...
swagger: "2.0"